requirements.txt
Flask
Flask-Cors
textblob
google-generativeai
gunicorn
